% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.pvs.R
\name{ind.pvs}
\alias{ind.pvs}
\title{A function to identify and extract the principal variables in your dataset.}
\usage{
ind.pvs(variabledata, tree, cormat, distmat, cutheight)
}
\arguments{
\item{variabledata}{a data.frame of the variable data used to build your cormat and hclust tree}

\item{tree}{a hclust object of you variables, built with the cormat and the hclust method "complete"}

\item{cormat}{your correlation matrix}

\item{distmat}{your distance matrix}

\item{cutheight}{the height at which you wish to cut your tree}
}
\description{
A function to identify and extract the principal variables in your dataset. The function will iterate over each newly generated list of PVs to insure that they do meet your tree cutheight standards.
}
\examples{
ind.pvs()
}
\keyword{cut}
\keyword{principal}
\keyword{tree}
\keyword{variables,}
