% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPVs.R
\name{iPVs}
\alias{iPVs}
\title{A super function that runs through all of the necessary staeps for you to identify Independent Principal Variables}
\usage{
iPVs(
  variabledata,
  cor_method = "spearman",
  dist_method = "R2",
  hclust_meth = "average",
  cutheight = 0.5
)
}
\arguments{
\item{variabledata}{a data.frame of the variable data used to build your cormat and hclust tree}

\item{cutheight}{tree cut height. Value is equal to a dissimilarity of 1 - Spearman's rho.}
}
\description{
A super function that runs through all of the necessary staeps for you to identify Independent Principal Variables
}
\examples{
iPVs()
}
\keyword{cut}
\keyword{data}
\keyword{independent}
\keyword{principal}
\keyword{reduction,}
\keyword{tree}
\keyword{variables,}
