% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcluster.groups.R
\name{Kcluster.groups}
\alias{Kcluster.groups}
\title{A function to identify all members of a cluster, generated from iterative tree cuts}
\usage{
Kcluster.groups(ind_pv_iterations)
}
\arguments{
\item{ks}{a list of k ( cutree() ) group identifiers for each iterative tree cut performed to generate a fully independent listof PVs.}

\item{pvs}{a vector of identified PVs (principal variables)}
}
\description{
A function to identify all members of a cluster, generated from iterative tree cuts
}
\examples{
Kcluster.groups()
}
\keyword{cut,}
\keyword{independent}
\keyword{principal}
\keyword{tree}
\keyword{variables}
\keyword{variables,}
