% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcluster_PVs.R
\name{Kcluster_PVs}
\alias{Kcluster_PVs}
\title{A function to identify the Principal Variable of a group}
\usage{
Kcluster_PVs(variabledata, Kmembers, myPVs)
}
\arguments{
\item{variabledata}{a data.frame of the variable data used to build your cormat and hclust tree}

\item{Kmembers}{a list, where each object in the list is a vector of variable IDs, identifying the members of a cluster or group.}

\item{myPVs}{a vector of my Principal Variable (myPVs) for each cluster}
}
\description{
A function to identify the Principal Variable of a group. The final step in the iPVs pipeling to identify the final PV to stand as the tagging and independent variable for the cluster of variables.
}
\examples{
Kcluster_PVs()
}
\keyword{cut,}
\keyword{iPV}
\keyword{principal}
\keyword{tree}
\keyword{variables,}
