% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.ab.data.Me.R
\name{pca.ab.data.Me}
\alias{pca.ab.data.Me}
\title{A superfunction to estimate correlations and relationships among variables in a data set.}
\usage{
pca.ab.data.Me(wdata)
}
\arguments{
\item{variabledata}{a vector of class categorical, for all elements tested, identifying a category | or bin | box that each element belong to}
}
\description{
This function estimates a Spearman's correlation matrix, an abs(distance matrix), a hclust tree, a PCA from the correlation matrix, and numerous estimates of the effective number of variables in a data set (Me)
}
\examples{
pca.ab.data.Me()
}
\keyword{Me}
\keyword{correaltion}
\keyword{distance}
\keyword{hclust,}
\keyword{matrix,}
