% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVA.R
\name{PVA}
\alias{PVA}
\title{A Function to extract principle variables of a data set}
\usage{
PVA(responses, data, nvarselect = NULL, p.variance = 1, include = NULL, ...)
}
\arguments{
\item{responses}{a vector of column names to run PVA on}

\item{data}{a matrix of data, variables are in columns}

\item{nvarselect}{the number of variables to return}

\item{p.variance}{the proportion of variance you want you principal variable to explain}

\item{include}{a matrix of data, variables are in columns}
}
\description{
This function estimates the correlation structure among two factors, using a chi-square test and the Crammer's V statsitic
}
\examples{
PVA()
}
\keyword{analysis}
\keyword{principle}
\keyword{variable}
