% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S22update.R
\name{S22update}
\alias{S22update}
\title{Function that adds a variable and calcullates the partial correlation matrix}
\usage{
S22update(k, i, S22, nvar)
}
\arguments{
\item{k}{i do not know what this is, it comes from the PV()}

\item{i}{i do not know what this is, it comes from the PV()}

\item{S22}{S22 to be updated}

\item{nvar}{the number fo variables}
}
\description{
This function estimates partial correlation matrix
}
\examples{
S22update()
}
\keyword{PVA}
