% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.builder.R
\name{tree.builder}
\alias{tree.builder}
\title{A superfunction to estimate correlations and relationships among variables in a data set.}
\usage{
tree.builder(
  wdata,
  cor_method = "spearman",
  dist_method = "R",
  hclust_meth = "complete",
  cmat = NULL
)
}
\arguments{
\item{wdata}{a data.frame of the variable data used to build your cormat and hclust tree, the variables should be in columns, samples in rows.}

\item{cor_method}{the correlation method used to construct the correlation matrix}

\item{dist_method}{the distance method to generate the distance matrix}

\item{hclust_meth}{the clustering method used by hclust}

\item{cmat}{a correlation matrix you can pass to the function, which will skip the generation of a new matrix.}
}
\description{
This function estimates a Spearman's correlation matrix, an abs(distance matrix), a hclust tree, a PCA from the correlation matrix, and numerous estimates of the effective number of variables in a data set (Me)
}
\examples{
tree.builder()
}
\keyword{Me}
\keyword{correaltion}
\keyword{distance}
\keyword{hclust,}
\keyword{matrix,}
