% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecut.iterator.4minrho.R
\name{treecut.iterator.4minrho}
\alias{treecut.iterator.4minrho}
\title{A function to iterate over various tree cut height and extract the minimum observed correaltion coefficient within a cluster.}
\usage{
treecut.iterator.4minrho(tree, cormat, cutheights)
}
\arguments{
\item{tree}{a hclust object of you variables, built with the cormat and the hclust method "complete"}

\item{cormat}{your data.frame of variables used to generate your tree}

\item{cutheights}{the heights at which you wish to cut your tree}
}
\description{
A function to iterate over various tree cut height and extract the minimum observed correaltion coefficient within a cluster.
}
\examples{
treecut.iterator.4minrho()
}
\keyword{coefficient}
\keyword{correlation}
\keyword{minimum}
\keyword{or}
\keyword{rho}
\keyword{treecut.sumstats,}
