% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecut.sumstats.R
\name{treecut.sumstats}
\alias{treecut.sumstats}
\title{A function to estimate some intra-cluster summary statstics in you hclust object.}
\usage{
treecut.sumstats(tree, cormat, cutheight)
}
\arguments{
\item{tree}{a hclust object of you variables, built with the cormat and the hclust method "complete"}

\item{cormat}{a correlation matrix among you variables}

\item{cutheight}{the height at which you wish to cut your tree}
}
\description{
This function estimates a few summary statistics (N, min, mean, max) for each cluster, with an N > 1, at your defined cutheight in your hclust object.
}
\examples{
treecut.sumstats()
}
\keyword{cut,}
\keyword{statistics}
\keyword{summary}
\keyword{tree}
