% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarRep.R
\name{VarRep}
\alias{VarRep}
\title{Variable Representation - PVA}
\usage{
VarRep(cormat)
}
\arguments{
\item{cormat}{your correlation matrix}
}
\description{
This function estimates how much of the total variance can be explained by each variable independently, in turn
}
\examples{
VarRep()
}
\keyword{PVA,}
\keyword{explained}
\keyword{principal}
\keyword{variables,}
\keyword{variance}
