% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztransform.R
\name{ztransform}
\alias{ztransform}
\title{A Function to Z transform a vector of data}
\usage{
ztransform(data)
}
\arguments{
\item{data}{the vector of data to be transformed}
}
\description{
A Function to Z transform a vector of data
}
\examples{
ztransform()
}
