% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPVs.R
\name{iPVs}
\alias{iPVs}
\title{A super function that runs through all of the necessary staeps for you to identify Independent Principal Variables}
\usage{
iPVs(
  variabledata,
  cor_method = "spearman",
  dist_method = "R",
  hclust_meth = "complete",
  cutheight = 0.5,
  cmat = NULL
)
}
\arguments{
\item{variabledata}{a data.frame of the variable data used to build your cormat and hclust tree}

\item{cor_method}{the correlation method used to construct the correlation matrix}

\item{dist_method}{the distance method to generate the distance matrix}

\item{hclust_meth}{the clustering method used by hclust}

\item{cutheight}{tree cut height. Value is equal to a dissimilarity of 1 - Spearman's rho.}

\item{cmat}{a correlation matrix you can pass to the function, which will skip the generation of a new matrix.}
}
\description{
A super function that runs through all of the necessary staeps for you to identify Independent Principal Variables
}
\examples{
iPVs()
}
\keyword{cut}
\keyword{data}
\keyword{independent}
\keyword{principal}
\keyword{reduction,}
\keyword{tree}
\keyword{variables,}
