% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.builder.R
\name{tree.builder}
\alias{tree.builder}
\title{A superfunction to estimate correlations and relationships among variables in a data set.}
\usage{
tree.builder(
  wdata,
  cor_method = "spearman",
  dist_method = "R2",
  hclust_meth = "average"
)
}
\arguments{
\item{variabledata}{a vector of class categorical, for all elements tested, identifying a category | or bin | box that each element belong to}
}
\description{
This function estimates a Spearman's correlation matrix, an abs(distance matrix), a hclust tree, a PCA from the correlation matrix, and numerous estimates of the effective number of variables in a data set (Me)
}
\examples{
tree.builder()
}
\keyword{Me}
\keyword{correaltion}
\keyword{distance}
\keyword{hclust,}
\keyword{matrix,}
