% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarRep.R
\name{VarRep}
\alias{VarRep}
\title{Variable Representation - PVA}
\usage{
VarRep(wdata, cor_method = "spearman")
}
\arguments{
\item{wdata}{a data frame with your variable in columns and samples in rows.}

\item{cor_method}{the type of regression to perform}
}
\description{
This function estimates how much of the total variance can be explained by each variable independently, in turn
}
\examples{
VarRep()
}
\keyword{PVA,}
\keyword{explained}
\keyword{principal}
\keyword{variables,}
\keyword{variance}
